rfsctl: support to remove filter paths by name
Hi Frantisek and all.

You will find attached a patch to add a new command to remove a path
from a filter by 'full path', following the advices you gave in a
previous mail.

I tested it under CentOS 5 and Ubuntu 11.04 and did not noticed any
problem.
Could you review it a tell me if you find something wrong (about locks,
resources, ...) ?

note 1: the code is mix between rfs_flt_paths_add() and
rfs_flt_paths_rem().
note 2: if a path is renamed, the command must a called with the new
name.
note 3: if the file/directory is deleted, it can not be removed by path,
even if a new file or directory is re-created with the same name.
note 4: the caller gets a ENOENT if the path given does not exist and a
EINVAL if the path given is not in the path list of the filter.
note 5: in the command line tool, I had to split the help in two strings
to avoid a 'string too long' warning.

Regards,
Loïc

PS: I send the patch with git-send-email. The command did not return any
error, but I not sure it was really sent. It may find its ways to the
list, some day ;-)

Signed-off-by: Frantisek Hrbata <fhrbata@redirfs.org>